component 5 trial 2
Trail 2 - I made a few adjustments to my code, and this trial
was a lot easier than the previous one since all I had to do
was introduce an if choice command so that if the user
input is 1. 2, or 3, the code accepts it, and if not, they
receive an error, which was exactly what I wanted and
worked perfectly.